refactor: move/rename CopyUrl to its own file
